archwiki: LocalSettings: Enable subpages in the main namespace
The default value for the main namespace is `false`, because Wikipedia
does not use subpages for categorization anymore. That is not the case
on ArchWiki, so there is no reason why the main namespace should behave
differently. The most visible changes are relative links using the
normal double square brackets notation and automatic breadcrumb links
appearing at the top of the subpage, linking to each parent page that
exists.

See [1] for discussion and [2] for upstream documentation of this
setting. Also see the related FS#39668 [3], which will be solved by this
patch.

[1] https://wiki.archlinux.org/index.php/Help_talk:Style#Slashes_in_titles
[2] https://www.mediawiki.org/wiki/Manual:$wgNamespacesWithSubpages
[3] https://bugs.archlinux.org/task/39668