common: Force netmask to /32
For some reason orion has a netmask of /8 which doesn't work in
hetzner's network since all traffic to other machines must pass through
a switch. Fix this by forcing a /32 netmask.

This change also requires us to set the gateway via the Peer option
because it would otherwise not be reachable.

Signed-off-by: Florian Pritz <bluewind@xinu.at>