roles/quassel: Speed up cleaning and increase verbosity
A full VACUUM takes quite a long time and requires an exclusive lock on
the table. Since our current quassel host is not starved for disk space,
use a normal VACUUM instead.

This speeds up cleaning from about 50s to about 6s. Most of that is the
CLUSTER, which also requires an exclusive lock. However, stalling the
quassel core for just 6 seconds every night is a lot less likely to
cause timeouts.