roles/flyspray: Setup directory handling
The setup directory for flyspray is present on our git, so, instead of being
deleted after the installation, it remains on the repository. To avoid issues
with it, it has permissions 000 when not in use. But, for cloning, it is
required to have write permissions. So, we do this permission juggling before
cloning.