common: Raise journald max files limit and disable error wall
Max files limit causes journald to be too agressive when cleaning up
old logs because it tries to stay under the limit. If there are enough
users on the system that won't happen and it will start removing system
logs way too soon (like after a few hours or days).

Signed-off-by: Florian Pritz <bluewind@xinu.at>