Add unbound
Spamassassin on apollo tries to connect to a local dns resolver, but
can't which leads to long timeouts. Flyspray sends mail via SMTP which
goes through SA and thus each request that sends a mail will also block
for a long time.

Fix this by adding unbound as a local resolver which is needed for
proper performance of SA since public resolvers are often blocked by
blacklists because they send too many requests.

Signed-off-by: Florian Pritz <bluewind@xinu.at>