syncrepo: Keep temporary files in dedicated directory
Without this hidden files will show up in the /srv/ftp tree and other
mirrors will sync these. Sometimes they get removed in the meantime and
then errors are printed on the downstream mirror. Just move them out of
the way and let rsync keep them in a dedicated directory until they are
ready to be moved into place.

Signed-off-by: Florian Pritz <bluewind@xinu.at>