added new variables to packer-io call
This will set the iso_check to none. We don't need that here hence
packer-io can't check GPG signatures.
We want to change iso_url to make sure that we use a local mirror of the
iso file. The {{isotime..}}-template is a variable that will change
every call of packer-io.