roles/archbuild: Use a more generic sudoers rule
Rather than explicitly listing each command allowed, match any
executable of the form /usr/bin/*-x86_64-build -- this will only ever be
a symlink to archbuild which is all we are looking for.

devtools allows new archbuild configurations to be easily added by
creating a symlink of the form ${config}-${arch}-build and adding the
relevant /usr/share/devtools/pacman-${config}.conf

This change allows the sudoers policy to automatically support these, in
the event that devtools distributes new archbuild configs or a sysadmin
adds a custom one.