archbuild: Disable units when archbuild_fs is not tmpfs
Installing some services and config files doesn't really matter. What
matters is if they are enabled so only use the conditional there. The
rest is easier and less confusing if it just installs the files all the
time. Otherwise we'd need a second task that performs the other case and
then if we change something we'd also have to adjust that task. That's
unnecessarily error prone.

Signed-off-by: Florian Pritz <bluewind@xinu.at>