common: Add syslog-ng and logrotate
journal logs are nice and all, but they are comparatively large and
(mostly) uncompressed. This way we at can at least maintain logs for a
reasonable time frame to investigate issues even if the journal decided
that it has to clean up "old" logs.

The logrotate config uses an archive directory to rotate logs to for
stricter permissions on old logs to increase data protection (EU GDPR).

Signed-off-by: Florian Pritz <bluewind@xinu.at>