archive: Fix redirect for packages with hyphen in pkgname
https://archive.archlinux.org/repos/2016/09/01/core/os/x86_64/iana-etc-20160513-1-any.pkg.tar.xz

should redirect to

https://archive.org/download/archlinux_pkg_iana-etc/iana-etc-20160513-1-any.pkg.tar.xz

but it actually redirects to the URL with the "-etc" part missing from
the directory name. The new regex fixes this by removing the part where
pkgver is allowed to have a hyphen. The PKGBUILD manpage explains that
pkgver is explicitly not allowed to contain a hyphen so we do not need
to allow this either.

It also makes the regex more readable by using named groups for the
different parts.

Signed-off-by: Florian Pritz <bluewind@xinu.at>