archive: Raise open file limit for uploader
stack pacakges currently hit the limit for some reason. There are ~1300
files to upload, but the default limit is 1024. The uploader throws an
exception and retires just to fail again. For now, just raise the limit
until the initial upload is over. Or leave it because it doesn't hurt.
Probably.

Signed-off-by: Florian Pritz <bluewind@xinu.at>