archive-uploader: Let systemd handle service restarts and use simple monitoring
It's actually quite difficult to check that a service has failed X times
within a time frame. Mostly because the service may be finished within a
few seconds and zabbix only fetches the state at one instance of time
and not the number of failures. To fix this, we go back to the old
monitoring, but let systemd restart the service.

Signed-off-by: Florian Pritz <bluewind@xinu.at>