zabbix: Let systemd.unit.is-enable always return a boolean
systemctl outputs an error on stderr if the unit does not exist. This
error is then combined with the other output in zabbix and results in an
unsupportd item, which makes it difficult to check that a unit is really
enabled. Since units that are not even installed are also not enabled,
we can safely ignore stderr here.

Signed-off-by: Florian Pritz <bluewind@xinu.at>