Add remaining PIA mirror boxes to also be archive-mirrors
This took quite a bit of manual work on the particular systems as they were partitioned in all sorts of ways that didn't allow for a big continuous chunk of storage such as required by the archive. I've manually repartitioned everything and it seems to work just fine. Now all systems have a big 4TB / and that's that.