roles/aurweb: Make the necessary changes to use memcached instead of apcu
Add memcached to the playbook and also change the php extensions to use memcached.
Removed the apcu options from defatuls and added memcached settings. Added the php-memcached
packages to the list of needed packages and also remove the apcu tasks. Added the memcached
systemd unit file and enable and start it from the tasks.