install_arch: Use compress-force=zstd instead of compress=zstd
Btrfs by default has a pretty stupid algorithm that checks for whether a file is
compressable or not. In practice this algorithm is pretty slow in my tests and
just directly passing the file to zstd is faster. I'm running compress-force=zstd
on all my machines now and it seems safe. Let's roll with this.