install_arch: Make sure hcloud-init is installed during packer run
Apparently before if we ran a packer build to generate the arch base image, the problem was that it wouldn't pick up the fact that the machine was an hcloud server as that would only be checked after the first boot due to an ordering issue.
Now we manually make sure that packer builds (which are always on hcloud anyway) will always have the hcloud-init files installed properly.