Add rebuilderd_results Prometheus metric
To monitor if reproducible builds are going in the right direction,
record the good/bad/unknown metrics from rebuilderd with a Prometheus
textcollector for a Grafana dashboard to display a long term trend.

A Python script is required to handle data collection as obtaining the
status with jq/bash is non trivial and cannot easily dnyamically collect
suites and statuses.

Closes: #146