roles/archweb: Add the service_nginx_template variable and fix when the maintenance mode should run
Configured the variable for the custom nginx template used on maintenance mode. It is important that this template handles the maintenance
on it's own. Also, the maintenance mode was running on gemini, even though archweb_site is false there. Add a check for archweb_site, to make
sure the maintenance mode only runs on the machine hosting the site.