tf-stage1/archlinux: Change DNS records for the archweb migration and also increase the machine size
Changed the A and AAAA record for @ to point to the new machine. Also, since we can't have CNAME to
@, change www from CNAME to A and AAAA and point to the new machine too. The other domains were kept
CNAME, but pointing to www in this case. Also, the master-key domain only had an A record, so added an
AAAA record too. The TTL is still at 600, but we can increase after a while.