Fix python module deployment issues
With creates: once the aur python module is initially deployed it and
it's console-scripts will never be updated. To update the
console-scripts when the git repository is updated or when a major
Python update is done. As figuring out if a major Python update is done
we simply check if the python module is importable.

Fixes: #164