dovecot: Only enable needed sieve extensions
The enotify extension can be used to send mails as other users, so let's
only enabled the sieve extensions, which we need.
The extensions was found by running `grep require` on the existing sieve
user scripts. mailbox wasn't in that list, but it is used by the global
spam-to-folder.sieve script.