archwiki: Decrease number of workers
The box now is much smaller than the box we had the wiki on before and so we used a lot of workers because it didn't matter.
The new box had too much memory pressure and so we had to decrease this.
In a load test, it lead to much better performance under high HTTP load.