Split out borg into two services
The reason is that we might want to independently retry the offsite backup in case it fails but not rerun the primary backup in that case.
This isn't possible with the current setup and so we'll have to split this up into two oneshot services.
The way it works is that borg-backup.service will schedule borg-backup-offsite.service to run after it which ensures that
no two backups are running at the same time and also that they can be indendently retried if one of them fails.