Add a trap to borg backup for cleanup
Now that our textcollector depends on the condition of /backup being
available, we should be extra careful that we get rid of the directory.
If anything in the script fails, a trap now umounts and removes the
directory automatically if on btrfs.being available, we should be extra
careful that we get rid of the directory. If anything in the script
fails, a trap now umounts and removes the directory automatically if on
btrfs.