rspamd: Lower spam threshold on misaligned Reply-To/To fields
When people send patches to pacman-dev, either with the wrong list
address or a perceived wrong email header it would most likely be
default marked as spam and stuffed into Junk for people using our email
server.

This attempts at lowering the score to something we can live with until
a gitlab migration.

Signed-off-by: Morten Linderud <morten@linderud.pw>