dovecot: Enable the variables Sieve extension
"In advanced mail filtering rule sets, it is useful to keep state or
configuration details across rules.  This document updates the Sieve
filtering language (RFC 5228) with an extension to support variables.
The extension changes the interpretation of strings, adds an action to
store data in variables, and supplies a new test so that the value of a
string can be examined."[1]

[1] https://datatracker.ietf.org/doc/html/rfc5229/

Fix #368