matrix: Use node --dns-result-order=ipv4first to work around bug
Node's address lookup strips the scope ID from link-local IPv6
addresses, making them invalid and unable to be connected to.

We're getting hit with this since Node 17 because it now defaults to the
`verbatim` dns-result-order and this happens to return the link-local
IPv6 address for `matrix.archlinux.org` first.

Specify `ipv4first` so we return to Node 16's behavior.