Provision shadow copy of repos for testing git packaging
For the upcoming "Git packaging sources"[1], we need a shadow copy of
repos.archlinux.org we can test stuff on (Ansible roles, packaging
workflow etc.). We still have repos-git.archlinux.org, but that box is
more of a POC box allocated to anthraxx for his MRs related to the git
migration.

More details on how the box was setup can be found in[2]. Only the
terraform code is committed to the master branch, as it is annoying
having the server in the state without the corresponding IaC code
comitted.

This was deployed Fri Dec 16 01:10:15 2022 +0100.

[1] https://lists.archlinux.org/archives/list/arch-dev-public@lists.archlinux.org/thread/YKKYBXQL62U5RTYIRI2NT2I3EG7V63HT/
[2] https://gitlab.archlinux.org/archlinux/infrastructure/-/merge_requests/665

Fix #479