archwiki: Update to 1.41.1-2
Version 1.41.1-2 adds support for the Timeless skin, see
https://gitlab.archlinux.org/archlinux/archwiki/-/merge_requests/71

The 'MonoBook' skin is replaced with 'Timeless' in the configuration
file. MonoBook is quite ugly and legacy skin that should be phased out.
When it is disabled, the skin-specific CSS overrides could be cleaned up
in the archwiki repository.

Also removed outdated comment referring to "internal symbolic names"
that are in a legacy section in the MediaWiki documentation:
https://www.mediawiki.org/wiki/Special:MyLanguage/Manual:$wgDefaultSkin#Legacy_information