install_arch: improve the ESP file system label
Use an uppercase LABEL to avoid a warning from mkfs.fat:

    mkfs.fat: Warning: lowercase labels might not work properly on some systems

Use something other than "efi" as the label since some firmware do not
like it. While this is not really an issue for Arch infrastructure, since
there is zero change that this will be run on any MSI hardware with its
broken firmware, it is best to simply folow the wiki anyway. See
https://wiki.archlinux.org/title/EFI_system_partition#Firmware_does_not_see_the_EFI_directory